feat: [PAYMCLOUD-109] POC Github Runner on DevOpsLab (#139)
* Add infrastructure setup for GitHub runner

This commit introduces new Terraform configurations to set up the infrastructure for a GitHub runner. It includes the addition of resource group, networking, container app environment and jobs, as well as corresponding variables and outputs. Additionally, a GitHub actions workflow for manual dispatch and supporting scripts have been added.

* Remove test-pipe workflow and add github-runner README

Deleted the manual dispatch workflow configuration from the test-pipe.yml file to streamline the repository. Added a new README.md for the GitHub runner in src/github-runner, providing detailed information on the requirements, modules, resources, inputs, and outputs for easy reference.

* Add manual dispatch workflow for self-hosted runner

This commit introduces a new GitHub Actions workflow that can be manually triggered via the workflow_dispatch event. The workflow includes steps to checkout the repository, echo a message, and run the `arp -a` command for testing purposes.

* Remove test-pipe workflow and update GitHub Runner config

Deleted the manual dispatch workflow YAML file and enhanced the GitHub Runner setup. Updated configurations to use specific branches and added managed identity outputs along with resource assignments.

* Update module source to tagged version v8.50.0

Changed the Terraform module source reference from a branch to a tagged version for improved stability and consistency. This ensures that the module uses a specific release version.